Add libgue TODO
The added TODO is a list of things known to be broken with regards to
libgue's parsing. I wrote them down rather than fixing them because I'm
lazy, but I _do_ plan to fix them in the future.

On the off chance someone is actually reading this commit message, PRs
welcome :)