libgue: Correctly handle different encodings
Cue sheets are often generated by Windows programs, and being a textual
format means they're vulnerable to being in a variety of different
encodings. The libgue scanner will choke on anything that isn't UTF-8,
so this commit adds code to detect the source charset and convert before
trying to parse.

This is a summary of changes:
 * Bump version to 0.2, since the libgue API has changed a little
 * Added dependency on uchardet + VAPI for internal use. README has been
    updated accordingly.
 * The main Gue.Sheet constructor has been changed to parse_data()
 * If the first couple of bytes passed to parse_data() are 0xef 0xbb
    0xbf (https://en.wikipedia.org/wiki/Byte_order_mark#UTF-8), those
    first three bytes are discarded from the buffer.
 * Project-wide warning squashing has been reinstated in configure.ac
 * One of the error messages in cue-munger.vala doesn't print \n after
    the error message; this commit fixes that too