libgue: Detect ExactAudioCopy cue sheets
ExactAudioCopy produces sheets that are not even slightly
spec-compliant.  Disabling known broken features is easier than trying
to make sense of broken sheets, so for the time being this is what will
happen. I do intend to fix this in the future, but this will do for now.

This adds a new bit of API: a bool flag, `generated_by_eac`, has been
added to `Gue.Sheet`. As such, version has been bumped to 0.6.

This commit also adds the first test! Hooray!