Try to test whether EAC cuesheet is malformed
I've just noticed a well-formed EAC cuesheet, the difference appearing
to be that this one had only a single file. Having a closer look at the
sample EAC file in the tests directory, it would seem as though the FILE
command is the main culprit. This commit tries to accommodate for this
in libgue.

This commit also bumps the version to 0.7.5.