Rewrite parser
This commit introduces a parser rewrite to go from a rather opaque
system of GLists containing indexes with special meaning to an AST-like
system with features like tree validation.

Although I've tried to make sure the API stayed as similar as possible,
I have had to shuffle some stuff around. As a result, I've bumped the
version number to 0.8

Some of the changes include:
 * Tree validation
 * Complete support for EAC-mangled sheets
 * Error messages now display the location of the problem in the sheet
 * Scanner shouldn't leak strings any more (though I haven't tested)